Update `Map` to not use `Rc` and `Cell` to hold the claimer.
`Rc` and `Cell` do not implement `Send|Sync`, which prevented us from having the overall `Map` shared across threads.

As the underlying type (`Option<usize>`) implements `Copy`, the claimer was rewritten as an `Arc` of an `AtomSetOnce` (the latter being a wrapper around an atomic pointer, with the added restriction that its value pointed to can only be set once).

Simultaneously, I changed `set_claimer` to take a `&mut self`, which is strictly not necessary but better documents what this does. In any case, only a mutable `Player` should end up modifying the claimer, so this breaks nothing downstream.